Make Zinc `Global` implementations independent
The callback implementation should be independent of the compiler
interface and the source code should reflect so.

Breaking change in comparison with the previous API: `Compiler` is now
called `ZincCompiler` to make the naming more clear and better
distinguish between `CachedCompiler` and its underlying `ZincCompiler`.

Docs have been added and simplifications to methods have been provided
and double checked with previous Scala versions. For instance, `dropRun`
has been confirmed to only exist in 2.10.x series and removed in 2.11.x.
Therefore, its implementation has been moved to the compatibility traits
for 2.10.x series.  We then remove a reflective call for versions above
2.10.6.
