+++ internal/compiler-bridge/src/main/scala_2.11+/xsbt/Compat.scala
@@ -1,8 +1,18 @@
 package xsbt
 
+import xsbti.compile.Output
+
+import scala.tools.nsc.Settings
+
 abstract class Compat
 object Compat
 
+/** Defines compatibility utils for [[ZincCompiler]]. */
+trait ZincGlobalCompat {
+  protected def superDropRun(): Unit = ()
+}
+
 private trait CachedCompilerCompat { self: CachedCompiler0 =>
-  def newCompiler: Compiler = new Compiler()
+  def newCompiler(settings: Settings, reporter: DelegatingReporter, output: Output): ZincCompiler =
+    new ZincCompiler(settings, reporter, output)
 }
